fix: Components fully responsive
