+++ src/components/JobCard/index.tsx
@@ -1,100 +1,67 @@
+import React from 'react';
+import { formatDistanceToNow } from 'date-fns';
 import {
-    Description,
-    Footer,
-    Header,
-    JobType,
-    Text,
+    Card,
+    Company,
+    Content,
+    Location,
+    Logo,
+    PublishedAt,
     Title,
-    Wrapper,
-    Button,
-    DateText,
-    PreviewField,
-    DescriptionPreview,
-    WrapperPreview,
-    Divider,
+    Type,
 } from './styles';
-import {
-    RiEditFill as PencilIcon,
-    RiDeleteBin6Fill as DeleteIcon,
-} from 'react-icons/ri';
-import { JobData } from '../../pages/InsertJobs';
-import format from 'date-fns/format';
+import { ptBR } from 'date-fns/locale';
 
-interface JobCardProps extends JobData {
-    onDeleteJobCard: (id: string) => void;
+export interface Props {
+    id: string;
+    title: string;
+    company: string;
+    city: string;
+    federalUnit: string;
+    jobType: string;
+    typeContract: string;
+    modality: string;
+    publishedAt: string;
+    onClick?: () => void;
+    active: boolean;
+    opacity: any;
 }
 
-export default function JobCard({
+const JobCard: React.FC<Props> = ({
     id,
     title,
-    description,
-    jobType,
-    createdAt,
-    onDeleteJobCard,
-    type,
-    salaryMin,
-    salaryMax,
-    prerequisites,
-    contractType,
-    benefits,
-    affirmative,
-    affirmativeType,
-    modality,
+    company,
     city,
-    uf,
-}: any) {
-    const NewDate: any = format(new Date(createdAt), 'dd/MM/yyyy - HH:mm');
+    federalUnit,
+    jobType,
+    typeContract,
+    modality,
+    publishedAt,
+    onClick,
+    active,
+    opacity,
+}) => {
+    const timeAgo = formatDistanceToNow(new Date(publishedAt), {
+        locale: ptBR,
+        addSuffix: true,
+    });
 
     return (
-        <WrapperPreview
-            id={id}
-            className="shadow-md shadow-[#1a1a1a]/5 hover:shadow-md"
-        >
-            <Header>
-                <div>
-                    <Title className="block">{title}</Title>
-                    <div className="flex items-center gap-[0.375rem]">
-                        <JobType className="text-[1.125rem]">{type}</JobType>
-                        <span className="h-1 w-1 bg-gray-dark/70 rounded-full" />
-                        <DateText>{NewDate}</DateText>
-                    </div>
-                </div>
-                <PencilIcon className="absolute right-12 hover:text-blue-500 text-[2.75rem] text-blue align-middle bg-blue-lighter border border-blue p-[0.4375rem] rounded" />
-            </Header>
-            <DescriptionPreview>
-                <Text dangerouslySetInnerHTML={{ __html: description }}></Text>
-            </DescriptionPreview>
-            <PreviewField>
-                Faixa Salarial: de {salaryMin} a {salaryMax}
-            </PreviewField>
-            <span>
-                <PreviewField>Pré-requisitos:</PreviewField>
-                <PreviewField>
-                    <p>{prerequisites}</p>
-                </PreviewField>
-            </span>
-            <PreviewField>{contractType}</PreviewField>
-            {benefits === null ? null : (
-                <PreviewField>Benefícios: {benefits}</PreviewField>
-            )}
-            {affirmative === 'false' ? null : (
-                <PreviewField>
-                    Grupo Minoritário: {affirmativeType}
-                </PreviewField>
-            )}
-            <PreviewField>
-                {modality !== 'Remoto'
-                    ? modality
-                    : `${modality} - ${city || ''}/${uf || ''}`}
-            </PreviewField>
-
-            <Footer>
-                <Divider />
-                <Button onClick={() => onDeleteJobCard(id)} type="button">
-                    <DeleteIcon className="text-base" />
-                    Remover
-                </Button>
-            </Footer>
-        </WrapperPreview>
+        <Card tabIndex={0} onClick={onClick} active={active} opacity={opacity}>
+            <Logo src="{companyLogoUrl}" alt={company} />
+            <Content>
+                <Title>{title}</Title>
+                <Company>{company}</Company>
+                <Location>
+                    {city}/{federalUnit}
+                </Location>
+                <Type>
+                    {modality} - {jobType} - {typeContract}
+                </Type>
+            </Content>
+            <PublishedAt>{timeAgo}</PublishedAt>
+        </Card>
     );
-}
+};
+
+export default JobCard;
