+++ src/components/JobDetails/CompanyLogo.tsx
