+++ src/pages/styles/FeedVagasStyles.ts
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { RiInformationLine } from 'react-icons/ri';
 import FeedJobsBackground from "../../assets/imgs/feedjobs-bg.svg"
 
 export const Wrapper = styled.div`
@@ -37,6 +36,8 @@ export const JobContainer = styled.div`
     align-items: center;
     margin-bottom: 56px;
     width: 100%;
+    padding-right: 20px;
+    padding-left: 20px;
 `
 
 export const PageTitle = styled.h1`
@@ -97,50 +98,6 @@ export const JobDetailsWrapper = styled.div`
   margin-top: 52px;
   display: flex;
   justify-content: center;
-
-`;
-
-export const NoJobsContainer = styled.div`
-  height: 569px;
-  max-width: 546px;
-  width: 100%;
-  border: 2px dashed #1165BA;
-  border-radius: 0.375rem;
-  text-align: center;
-  padding: 6rem;
-`;
-
-export const InnerContainer = styled.div`
-  margin-bottom: 45px;
-  margin: 0 auto;
-  border: 1px solid #1165BA;
-  border-radius: 50%;
-  width: 10.625rem;
-  padding: 36px;
-`;
-
-export const Icon = styled(RiInformationLine)`
-  font-size: 6rem;
-  color: #1165BA;
-`;
-
-export const Title = styled.p`
-  margin-top: 45px;
-  font-size: 25px;
-  line-height: 30px;
-  font-weight: 400;
-  margin-bottom: 0.4375rem;
-  font-family: Radio Canada;
-  color: #515050;
-  margin-bottom: 7px;
-`;
-
-export const Message = styled.p`
-  font-family: Radio Canada;
-  font-size: 23px;
-  line-height: 28px;
-  font-weight: normal;
-  color: #1165BA;
 `;
 
 export const ShowMore = styled.button `
@@ -149,6 +106,10 @@ export const ShowMore = styled.button `
   color: #515050;
   grid-column-start: 1;
   margin-top: 20px;
+
+  :hover {
+    text-decoration: underline;
+  }
 `;
 
 export const NoResultsMessage = styled.div`
